Fixes warnings/errors for 32-bit (and maybe ILP64?) builds
BugFix: ARM32, and possibly other archs, fails with -Werror

The code implicitly casts a uint64_t to a vector::size_type. It does so only after checking bounds (against a value that must have come from a size_type, and without using signed types anywhere), so it should be safe, but it still causes warnings.